Example hand: unseated player taking actions
At PartyPoker if a player sits down to a table after a small blind has
been posted but every possible big blinds still not posted (e.g. by new
players at the table) then it is possible to not have the player's name
in the seat list but he/she can take actions during the hand.