Remove trailing percent symbols from HUD values
The values that are displayed in the HUD box included the '%' symbol if
these values were percentages. This clutters the HUD for those who don't
need such extra characters.

This patch removes the trailing '%' from result[1]. The default
formatting for HUD values is '%3.1f' and if decimal places need to be
changed, the code for those is now simpler. A user who wishes to show a
percent symbol after any given stat value, can edit HUD_config.xml and
assign such stats the hudsuffix="%" attribute.

This patch also changes the way the tooltip strings are constructed.
Instead of appending the percent symbol to generated string, the '%' is
now generated in place by the format-string stanza '%%'.