Fix HUD autoclose for wnck Xtables
Add code to track the live table windows with Wnck. The toolkit allows
to monitor events of closed windows, but only on global level.
Connecting to individual windows' termination events is not possible.

Add a module-global cache of known table windows, with their XID values
as keys. Each time a window is closed, compare the XID and remove from
cache if it matched a known table window.

With this fix, closing a poker table will automatically close the
associated HUD as well.