Add code for loading a whole SVG deck
This introduces a new module, Deck.py which is used to load a complete
deck of SVG cards, and to access the cards as images. The implementation
has two purposes:
    1) to allow changing the deck as simply as possible
    2) to make it easy to get the image for any given card

Card images are placed in a dictionary so that any given card can be
looked up by Deck.card(suit, rank)

Path to deck directory is looked up from card_path.py which is only used
by the deck module. The file can be automatically generated for packaged
installations so that deck paths may be treated as always known and not
cared about in actual FPDB code.