Fix mucked card display for SVG deck
This patch reworks the logic behind "seats with known cards" and "known
cards for seat".

We have already split the uses of has_cards(), but the logic how it was
used was somewhat hazy. To make code more maintainable, we split the
functionality. The old routine is now explicitly used to fetch the
number and location of seats that showed cards. The newly introduced
valid_cards() is then used to figure out exactly how many cards a seat
showed. In case of community cards, board is treated as a seat as
before.

The split is necessary to make the code more readable. Return value of
valid_cards() is used to allocate the pixbuf for given seat. Since each
seat always has a 7-tuple to announce the cards, we have to check how
many valid cards there were.